adjusting scripts

+++ experiments/exp2.sh
@@ -9,15 +9,15 @@ mkdir -p logs_${LOG_TIMESTAMP}
 cp experiments/kill_processes.py logs_${LOG_TIMESTAMP}
 
 echo "Start server broker"
-docker stop server-broker-rabbit
-docker rm server-broker-rabbit
-docker run --hostname broker --name server-broker-rabbit -p 9000:5672 rabbitmq:3 > logs_${LOG_TIMESTAMP}/cloud_broker.log 2>&1 &
+sudo docker stop server-broker-rabbit
+sudo docker rm server-broker-rabbit
+sudo docker run --hostname broker --name server-broker-rabbit -p 9000:5672 rabbitmq:3 > logs_${LOG_TIMESTAMP}/cloud_broker.log 2>&1 &
 sleep 5
 
 echo "Start client broker"
-docker stop client-broker-rabbit
-docker rm client-broker-rabbit
-docker run --hostname broker --name client-broker-rabbit -p 8000:5672 rabbitmq:3 > logs_${LOG_TIMESTAMP}/client_broker.log 2>&1 &
+sudo docker stop client-broker-rabbit
+sudo docker rm client-broker-rabbit
+sudo docker run --hostname broker --name client-broker-rabbit -p 8000:5672 rabbitmq:3 > logs_${LOG_TIMESTAMP}/client_broker.log 2>&1 &
 sleep 10
 
 echo "Start event reader"
@@ -45,7 +45,7 @@ echo $! >>  logs_${LOG_TIMESTAMP}/pids
 sleep 1
 
 echo "Start user manager"
-rm user_manager/db/users.db
+[ -d "user_manager/db/users.db" ] && /rm user_manager/db/users.db
 python -u -m user_manager.service_user_manager > logs_${LOG_TIMESTAMP}/service_user_manager.log 2>&1 &
 echo $! >>  logs_${LOG_TIMESTAMP}/pids
 sleep 1
@@ -66,8 +66,8 @@ echo $! >>  logs_${LOG_TIMESTAMP}/pids
 sleep 1
 
 echo "Start client tasks manager" 
-rm -rf client_task_manager/tasks/*
-rm -rf client_task_manager/client_info/*
+[ -d "client_task_manager/tasks/" ] && rm -rf client_task_manager/tasks/*
+[ -d "client_task_manager/client_info/" ] && rm -rf client_task_manager/client_info/*
 python -u -m client_task_manager.service_client_ml > logs_${LOG_TIMESTAMP}/service_client_ml.log 2>&1 &
 echo $! >>  logs_${LOG_TIMESTAMP}/pids
 sleep 30
@@ -81,8 +81,8 @@ python kill_processes.py
 cd ..
 cp events.json logs_${LOG_TIMESTAMP}
 cp events.json experiments
-rm -r cloud_task_manager/tasks/task_E
-rm -r cloud_task_manager/tasks/task_C
+[ -d "cloud_task_manager/tasks/task_E" ] && rm -r cloud_task_manager/tasks/task_E
+[ -d "cloud_task_manager/tasks/task_C" ] && rm -r cloud_task_manager/tasks/task_C
 
 cd experiments
 echo "---- RESULTS ----" >> exp2_raw_times
